2008-11-21  Kai Tietz  <kai.tietz@onevision.com>
        *  Object.m (errno): Replaced by errno.h include.
        (compare): Cast self to id to prevent warning on comparison.
        * objc/objc.h (BOOL): Prevent redeclaration of BOOL, if it is
        already there.
        * sendmsg.c (__objc_print_dtable_stats): Remove type warnings.
        * thr-win32.c (__objc_thread_detach): Remove type warning.
        (__objc_thread_id): Likewise.
        * thr.c (__objc_thread_detach_functiont): Add __builtin_trap ()
        for noreturn.



git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@142087 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libobjc/thr-win32.c
@@ -70,7 +70,7 @@ __objc_thread_detach(void (*func)(void *arg), void *arg)
                                    arg, 0, &thread_id)))
     thread_id = 0;
   
-  return (objc_thread_t)thread_id;
+  return (objc_thread_t)(size_t) thread_id;
 }
 
 /* Set the current thread's priority. */
@@ -151,7 +151,7 @@ __objc_thread_exit(void)
 objc_thread_t
 __objc_thread_id(void)
 {
-  return (objc_thread_t)GetCurrentThreadId();
+  return (objc_thread_t)(size_t) GetCurrentThreadId();
 }
 
 /* Sets the thread's local storage pointer. */
